Created My own Simple Project
This React project is a bug tracker, which allows users to create, view, and update a list of bugs that need to be fixed. The app displays a table of bugs, each with a title, author, and fixed status. Users can delete a bug by clicking a button, and they can also filter the list to show only bugs that have been fixed. :smirk: